Move context parallelism code into it's own file. Modify test and local imports of cp code accordingly
